[HWKMETRICS-523] Relocate the original tag parser to the new generic tag package.
