+++ core/metrics-core-service/src/main/java/org/hawkular/metrics/core/service/tags/SimpleTagQueryParser.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2016 Red Hat, Inc. and/or its affiliates
+ * Copyright 2014-2017 Red Hat, Inc. and/or its affiliates
  * and other contributors as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.core.service;
+package org.hawkular.metrics.core.service.tags;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -26,6 +26,9 @@ import java.util.Set;
 import java.util.TreeMap;
 import java.util.regex.Pattern;
 
+import org.hawkular.metrics.core.service.DataAccess;
+import org.hawkular.metrics.core.service.MetricsService;
+import org.hawkular.metrics.core.service.PatternUtil;
 import org.hawkular.metrics.core.service.transformers.ItemsToSetTransformer;
 import org.hawkular.metrics.core.service.transformers.TagsIndexRowTransformer;
 import org.hawkular.metrics.model.Metric;
@@ -41,12 +44,12 @@ import rx.functions.Func1;
  *
  * @author Michael Burman
  */
-public class TagQueryParser {
+public class SimpleTagQueryParser {
 
     private DataAccess dataAccess;
     private MetricsService metricsService;
 
-    public TagQueryParser(DataAccess dataAccess, MetricsService metricsService) {
+    public SimpleTagQueryParser(DataAccess dataAccess, MetricsService metricsService) {
         this.dataAccess = dataAccess;
         this.metricsService = metricsService;
     }
